requirements.txt
torch>=0.4.0
torchvision>=0.2.1
dominate>=2.3.1
numpy>=1.13.3
scipy==1.2.2
h5py>=2.7.1
pillow>=3.1.1
tqdm>=4.31.1
